refactor: error and greeting
