+++ src/config/index.js
@@ -1,0 +1,5 @@
+const config = require('./config');
+
+module.exports = {
+  ...config,
+};
