--- src/display/index.js
@@ -1,7 +1,0 @@
-const greeting = require('./greeting');
-const error = require('./error');
-
-module.exports = {
-  ...greeting,
-  ...error,
-};
