+++ src/error/index.js
@@ -1,0 +1,9 @@
+const errors = require('./errors');
+const { createErrorText } = require('./createErrorText');
+const { showError } = require('./showError');
+
+module.exports = {
+  ...errors,
+  createErrorText,
+  showError,
+};
