--- src/errors/index.js
@@ -1,3 +1,0 @@
-const errors = require('./errors');
-
-module.exports = errors;
