+++ src/greeting/index.js
@@ -1,0 +1,5 @@
+const greeting = require('./showGreeting');
+
+module.exports = {
+  ...greeting,
+};
